+++ src/main/java/com/victorseger/svtech/services/EstadoService.java
@@ -17,10 +17,13 @@ public class EstadoService {
     private EstadoRepository estadoRepository;
 
     public Estado find(Integer id) {
-        Optional<Estado> obj = estadoRepository.findById(id);
+        if (id != null) {
+            Optional<Estado> obj = estadoRepository.findById(id);
 
-        return obj.orElseThrow(() -> new ObjectNotFoundException(
-                "Objeto não encontrado! Id: " + ", Tipo: " + Estado.class.getName()));
+            return obj.orElseThrow(() -> new ObjectNotFoundException(
+                    "Objeto não encontrado! Id: " + ", Tipo: " + Estado.class.getName()));
+        }
+        return null;
     }
 
     public List<Estado> findAll() {
@@ -28,34 +31,42 @@ public class EstadoService {
     }
 
     public Estado getOne(Integer id) {
-        return estadoRepository.getOne(id);
+        if (id != null) {
+            return estadoRepository.getOne(id);
+        }
+        return null;
     }
 
     public Estado insert(Estado estado) {
-        if (estado.getNome() != null && !estado.getNome().isEmpty()) {
+        if (estado != null && estado.getNome() != null && !estado.getNome().isEmpty()) {
             estado.setId(null);
             return estadoRepository.save(estado);
-        } else return null;
+        }
+        return null;
     }
 
     public Estado update(Estado estado) {
-        if (estado.getNome() != null && !estado.getNome().isEmpty()) {
+        if (estado != null && estado.getNome() != null && !estado.getNome().isEmpty()) {
             Estado newEstado = find(estado.getId());
             //chama o método auxiliar para apenas atualizar os campos desejados do estado e não remover nenhum valor de outro campo
             updateData(newEstado, estado);
             return estadoRepository.save(newEstado);
-        } else return null;
+        }
+        return null;
     }
 
     public boolean delete(Integer id) {
-        boolean flag = true;
-        find(id);
-        try {
-            estadoRepository.deleteById(id);
-        } catch (DataIntegrityViolationException e) {
-            flag = false;
+        if (id != null) {
+            boolean flag=true;
+            find(id);
+            try {
+                estadoRepository.deleteById(id);
+            } catch (DataIntegrityViolationException e) {
+                flag = false;
+            }
+            return flag;
         }
-        return flag;
+        return false;
     }
 
     private void updateData(Estado newEstado, Estado estado) {
